THFREE-38. Change structure of project
